+++ build_tests/shared_library/shared_library_d_ldc2.rb
@@ -1,0 +1,7 @@
+configure do
+  check_d_compiler "ldc2"
+end
+
+env(echo: :command) do |env|
+  env.SharedLibrary("mine", glob("src/lib/*.d"))
+end